use next.js for player-example
