--- packages/player-example/config/webpack/persistentCache/createEnvironmentHash.js
@@ -1,9 +1,0 @@
-'use strict';
-const { createHash } = require('crypto');
-
-module.exports = env => {
-  const hash = createHash('md5');
-  hash.update(JSON.stringify(env));
-
-  return hash.digest('hex');
-};
