+++ packages/player-example/package.json
@@ -8,45 +8,23 @@
 		"Shankhadeep Dey <shankhadeepdey99@gmail.com>"
 	],
 	"dependencies": {
-		"@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
 		"@remotion/bundler": "3.3.33",
 		"@remotion/eslint-config": "3.3.33",
 		"@remotion/gif": "3.3.33",
 		"@remotion/player": "3.3.33",
 		"@remotion/preload": "3.3.33",
-		"@types/node": "^12.0.0",
 		"@types/react": "18.0.26",
 		"@types/react-dom": "18.0.10",
-		"browserslist": "^4.21.4",
-		"camelcase": "^6.2.1",
-		"esbuild": "0.16.12",
 		"eslint": "^8.3.0",
-		"fs-extra": "^10.0.0",
-		"html-webpack-plugin": "^5.5.0",
-		"react": "18.0.0",
-		"react-dev-utils": "^12.0.1",
-		"react-dom": "18.0.0",
-		"react-refresh": "^0.11.0",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
 		"remotion": "3.3.33",
 		"typescript": "^4.7.0",
-		"webpack": "5.74.0",
-		"webpack-dev-server": "^4.6.0"
+		"next": "^13.1.3"
 	},
 	"scripts": {
 		"lint": "eslint src --ext ts,tsx",
-		"start": "node scripts/start",
-		"build-site": "node scripts/build"
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
+		"dev": "next dev",
+		"build-site": "next build"
 	}
 }
