+++ packages/player-example/pages/_app.tsx
@@ -1,0 +1,7 @@
+import type {AppProps} from 'next/app';
+
+function MyApp({Component, pageProps}: AppProps) {
+	return <Component {...pageProps} />;
+}
+
+export default MyApp;
