--- packages/player-example/scripts/start.js
@@ -1,169 +1,0 @@
-'use strict';
-
-// Do this as the first thing so that any code reading it knows the right env.
-process.env.BABEL_ENV = 'development';
-process.env.NODE_ENV = 'development';
-
-// Makes the script crash on unhandled rejections instead of silently
-// ignoring them. In the future, promise rejections that are not handled will
-// terminate the Node.js process with a non-zero exit code.
-process.on('unhandledRejection', (err) => {
-	throw err;
-});
-
-const fs = require('fs');
-const chalk = require('react-dev-utils/chalk');
-const webpack = require('webpack');
-const WebpackDevServer = require('webpack-dev-server');
-const clearConsole = require('react-dev-utils/clearConsole');
-const checkRequiredFiles = require('react-dev-utils/checkRequiredFiles');
-const {
-	choosePort,
-	createCompiler,
-	prepareProxy,
-	prepareUrls,
-} = require('react-dev-utils/WebpackDevServerUtils');
-const openBrowser = require('react-dev-utils/openBrowser');
-const paths = require('../config/paths');
-const configFactory = require('../config/webpack.config');
-const createDevServerConfig = require('../config/webpackDevServer.config');
-const react = require(require.resolve('react', {paths: [paths.appPath]}));
-const publicUrl = paths.publicUrlOrPath.slice(0, -1);
-const env = {
-	// Useful for determining whether we’re running in production mode.
-	// Most importantly, it switches React into the correct mode.
-	NODE_ENV: process.env.NODE_ENV || 'development',
-	// Useful for resolving the correct path to static assets in `public`.
-	// For example, <img src={process.env.PUBLIC_URL + '/img/logo.png'} />.
-	// This should only be used as an escape hatch. Normally you would put
-	// images into the `src` and `import` them in code to get their paths.
-	PUBLIC_URL: publicUrl,
-	// We support configuring the sockjs pathname during development.
-	// These settings let a developer run multiple simultaneous projects.
-	// They are used as the connection `hostname`, `pathname` and `port`
-	// in webpackHotDevClient. They are used as the `sockHost`, `sockPath`
-	// and `sockPort` options in webpack-dev-server.
-	WDS_SOCKET_HOST: process.env.WDS_SOCKET_HOST,
-	WDS_SOCKET_PATH: process.env.WDS_SOCKET_PATH,
-	WDS_SOCKET_PORT: process.env.WDS_SOCKET_PORT,
-	// Whether or not react-refresh is enabled.
-	// It is defined here so it is available in the webpackHotDevClient.
-	FAST_REFRESH: process.env.FAST_REFRESH !== 'false',
-};
-const useYarn = fs.existsSync(paths.yarnLockFile);
-const isInteractive = process.stdout.isTTY;
-
-// Warn and crash if required files are missing
-if (!checkRequiredFiles([paths.appHtml, paths.appIndexJs])) {
-	process.exit(1);
-}
-
-// Tools like Cloud9 rely on this.
-const DEFAULT_PORT = parseInt(process.env.PORT, 10) || 3000;
-const HOST = process.env.HOST || '0.0.0.0';
-
-if (process.env.HOST) {
-	console.log(
-		chalk.cyan(
-			`Attempting to bind to HOST environment variable: ${chalk.yellow(
-				chalk.bold(process.env.HOST)
-			)}`
-		)
-	);
-	console.log(
-		`If this was unintentional, check that you haven't mistakenly set it in your shell.`
-	);
-	console.log(
-		`Learn more here: ${chalk.yellow('https://cra.link/advanced-config')}`
-	);
-	console.log();
-}
-
-// We require that you explicitly set browsers and do not fall back to
-// browserslist defaults.
-const {checkBrowsers} = require('react-dev-utils/browsersHelper');
-checkBrowsers(paths.appPath, isInteractive)
-	.then(() => {
-		// We attempt to use the default port but if it is busy, we offer the user to
-		// run on a different port. `choosePort()` Promise resolves to the next free port.
-		return choosePort(HOST, DEFAULT_PORT);
-	})
-	.then((port) => {
-		if (port == null) {
-			// We have not found a port.
-			return;
-		}
-
-		const config = configFactory('development');
-		const protocol = process.env.HTTPS === 'true' ? 'https' : 'http';
-		const appName = require(paths.appPackageJson).name;
-
-		const useTypeScript = fs.existsSync(paths.appTsConfig);
-		const urls = prepareUrls(
-			protocol,
-			HOST,
-			port,
-			paths.publicUrlOrPath.slice(0, -1)
-		);
-		// Create a webpack compiler that is configured with custom messages.
-		const compiler = createCompiler({
-			appName,
-			config,
-			urls,
-			useYarn,
-			useTypeScript,
-			webpack,
-		});
-		// Load proxy config
-		const proxySetting = require(paths.appPackageJson).proxy;
-		const proxyConfig = prepareProxy(
-			proxySetting,
-			paths.appPublic,
-			paths.publicUrlOrPath
-		);
-		// Serve webpack assets generated by the compiler over a web server.
-		const serverConfig = {
-			...createDevServerConfig(proxyConfig, urls.lanUrlForConfig),
-			host: HOST,
-			port,
-		};
-		const devServer = new WebpackDevServer(serverConfig, compiler);
-		// Launch WebpackDevServer.
-		devServer.startCallback(() => {
-			if (isInteractive) {
-				clearConsole();
-			}
-
-			if (env.FAST_REFRESH) {
-				console.log(
-					chalk.yellow(
-						`Fast Refresh requires React 16.10 or higher. You are using React ${react.version}.`
-					)
-				);
-			}
-
-			console.log(chalk.cyan('Starting the development server...\n'));
-			openBrowser(urls.localUrlForBrowser);
-		});
-
-		['SIGINT', 'SIGTERM'].forEach((sig) => {
-			process.on(sig, () => {
-				devServer.close();
-				process.exit();
-			});
-		});
-
-		if (process.env.CI !== 'true') {
-			// Gracefully exit when stdin ends
-			process.stdin.on('end', () => {
-				devServer.close();
-				process.exit();
-			});
-		}
-	})
-	.catch((err) => {
-		if (err && err.message) {
-			console.log(err.message);
-		}
-		process.exit(1);
-	});
