+++ packages/player-example/tsconfig.json
@@ -1,21 +1,37 @@
 {
-	"include": ["./src/**/*"],
-	"compilerOptions": {
-		"strict": true,
-		"esModuleInterop": true,
-		"lib": ["dom", "es2015"],
-		"jsx": "react-jsx",
-		"target": "es5",
-		"allowJs": true,
-		"skipLibCheck": true,
-		"allowSyntheticDefaultImports": true,
-		"forceConsistentCasingInFileNames": true,
-		"noFallthroughCasesInSwitch": true,
-		"module": "esnext",
-		"moduleResolution": "node",
-		"resolveJsonModule": true,
-		"isolatedModules": true,
-		"noEmit": true
-	},
-	"references": [{"path": "../player"}, {"path": "../core"}]
+  "include": [
+    "./src/**/*"
+  ],
+  "compilerOptions": {
+    "strict": true,
+    "esModuleInterop": true,
+    "lib": [
+      "dom",
+      "es2015"
+    ],
+    "jsx": "preserve",
+    "target": "es5",
+    "allowJs": true,
+    "skipLibCheck": true,
+    "allowSyntheticDefaultImports": true,
+    "forceConsistentCasingInFileNames": true,
+    "noFallthroughCasesInSwitch": true,
+    "module": "esnext",
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "incremental": true
+  },
+  "references": [
+    {
+      "path": "../player"
+    },
+    {
+      "path": "../core"
+    }
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
